# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Anatolia B.I.S. API",
    "version": "1.0.0",
    "description": "Machine-readable companion to API.md. See API.md for full prose documentation (request/response shapes, rate limits, error codes) — this file exists primarily so server/tests/openapi_drift.rs can catch a route that was added to, or removed from, server/src/routes/mod.rs without a matching documentation update. Kept intentionally lightweight rather than exhaustively typed."
  },
  "paths": {
...
    "/api/v1/admin/memberships": {
      "post": {
        "summary": "Assign a user to an organization/unit (SYSTEM_ADMIN only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "summary": "Remove a user's membership in an organization (SYSTEM_ADMIN only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/candidates/{candidateId}/entity-graph": {
      "get": {
        "summary": "A candidate's entity graph: alias/username/organization/website relations, organization-scoped",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden (role check, or organization scoping)"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "Manually record an alias/username/organization/website relation for a candidate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Validation error (unknown relation type, or empty value)"
          },
          "403": {
            "description": "Forbidden (role check, or organization scoping)"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}